float fix

+++ programs/distance_matrix.cl
@@ -4,7 +4,7 @@ __kernel void compute_distance_matrix(const int N, __global float *matrix, __glo
   int gid = get_global_id(0);
   int a_index = (int)(gid/N);
   int b_index = (int)(gid - (N * a_index));
-  double sum = 0.0;
+  float sum = 0.0;
 
   for(int i=0; i < size_v; i++){
     sum += pown(data[b_index][i] - data[a_index][i], 2);
